requirements.txt
python-telegram-bot==20.7
httpx==0.27.0
aiohttp==3.9.5
asyncio
